refactor: extract components
